Bug 1835241 - Part 4: Make narrowing conversions explicit in GC code r=sfink
This is the most complex of these patches. I've tried to simplify existing
code and make conversions explicit. In some places I've changed types so they
make more sense (e.g. HeapThreshold::eagerAllocTrigger now returns an integer
number of bytes). I've changed some uses of float to double as it's
more standard.

I've changed some of the telemetry helper classes too. Enumeration is now
unsigned and Percentage now a double. This fits better with their uses.

Differential Revision: https://phabricator.services.mozilla.com/D181509

+++ js/src/gc/MallocedBlockCache.h
@@ -77,7 +77,7 @@ class MallocedBlockCache {
   // Allows users to gradually hand blocks back to js_free, so as to avoid
   // space leaks in long-running scenarios.  The specified percentage of
   // blocks in each list is discarded.
-  void preen(float percentOfBlocksToDiscard);
+  void preen(double percentOfBlocksToDiscard);
 
   // Return all blocks in the cache to js_free.
   void clear();
